feat: Add hero video dialog component
The hero video dialog component has been added to the project. This component allows for the display of a video thumbnail with a play button, and opens a dialog with the full video when clicked. It supports various animation styles and can be used to showcase videos on the landing page and other sections of the application.